Added pause button and auto-turret v1.
Added constructors to make buttons (e.g. a button to temporarily deactivate auto-turrets), and a pause-button made with it.
Added a constructor to make auto-turrets, as well as an auto turret (for testing) placed in the corner of the screen.
Currently still working on targeting algorithm for the auto-turrets, ideally I want an algorithm that lets them perfectly
hit any target they aim at (by calculating the necessary lead), then we can worry about balancing it later.